Upgrade to Hibernate 3.2
Update unit tests to pass using 3.2 as there were some changes in 3.2 as to when
session state was flushed to db.  Verified that functional tests pass.  Now that
we are using 3.2 we can look into using annotations and the validator framework.

git-svn-id: svn+ssh://svn.osafoundation.org/svn/server/cosmo/trunk@2725 5fd0b11a-f2f7-0310-948e-c8f80ebed640

+++ src/main/java/org/osaf/cosmo/dao/hibernate/UserDaoImpl.java
@@ -119,10 +119,10 @@ public class UserDaoImpl extends HibernateDaoSupport implements UserDao {
             List results = crit.list();
 
             // Need the total
-            Integer size = (Integer) getSession().createQuery(
+            Long size = (Long) getSession().createQuery(
                     "select count(*) from User").uniqueResult();
 
-            return new ArrayPagedList(pageCriteria, results, size);
+            return new ArrayPagedList(pageCriteria, results, size.intValue());
         } catch (HibernateException e) {
             throw SessionFactoryUtils.convertHibernateAccessException(e);
         }
